Update to use php-solr-client library without relocation of same, and include the dependent library directly in the module.
Per ISLANDORA-233, dealing with the Solr directory (and the contained php-solr-client library) is a  development hassle.  It's also a distribution hassle, as the third-party library needed to be gotten after the fact.  

My hope is that, going forward, we won't need to think about the php-solr-client library either when developing or when deploying.

php-solr-client copyright notice states that modification/redistribution are allowed provided copyright notice is retained intact.  

Previously I (Zac) removed the php-solr-client library from our source tree because I wasn't sure we could redistribute it.  In fact, we had been throwing away everything but the library itself -- including the copyright notice -- so we were indeed in violation.  This shouuld be kosher.