Remove t() calls where inappropriate.
The t() function should be thought of as standing for "translate" not "text" as is commonly thought.  A good rule of thumb is if it's not potentially translatable, there is no sense wrapping it in a t() call.  We have a number of uses of t() on empty strings (which isn't harmful but offers no benefit), but also a number of calls to t() on configuration variables.  Users could potentially translate these into local languages and the result would be to break configurations, possibly making configuration impossible.