Improve handling of pluses ("+") in query strings.
For reasons unknown to me, drupal-made links with pluses in them that are part of an item list come out clean, but those which are outside of an item list do standard, annoying, plus-to-space conversion in the href of the resulting <a> tag.  I've modified the slash-replacement and restoration functions to include funky strings for pluses also.